backend(server): move language_rank implementation
