--- atcoder-problems-backend/src/server/language_count_ranking.rs
@@ -1,34 +1,0 @@
-use crate::server::{AppData, CommonResponse};
-use serde::{Deserialize, Serialize};
-use sql_client::language_count::LanguageCountClient;
-use tide::{Request, Response, Result};
-
-pub(crate) async fn get_users_language_rank<A>(
-    request: Request<AppData<A>>,
-) -> Result<Response> {
-    #[derive(Debug, Deserialize)]
-    struct Query {
-        user: String,
-    }
-    #[derive(Debug, Serialize)]
-    struct UsersLanguageResponse {
-        language: String,
-        count: i64,
-        rank: i64,
-    }
-    let conn = request.state().pg_pool.clone();
-    let query = request.query::<Query>()?;
-    let counts = conn.load_users_language_count(&query.user).await?;
-    let ranks = conn.load_users_language_count_rank(&query.user).await?;
-    let info = counts
-        .into_iter()
-        .zip(ranks)
-        .map(|(c, r)| UsersLanguageResponse {
-            language: c.simplified_language,
-            count: c.problem_count as i64,
-            rank: r.rank,
-        })
-        .collect::<Vec<_>>();
-    let response = Response::json(&info)?;
-    Ok(response)
-}
