+++ atcoder-problems-backend/src/server/mod.rs
@@ -1,6 +1,5 @@
 pub(crate) mod auth;
 pub(crate) mod internal_user;
-pub(crate) mod language_count_ranking;
 pub(crate) mod middleware;
 pub(crate) mod problem_list;
 pub(crate) mod progress_reset;
@@ -15,10 +14,9 @@ pub(crate) mod virtual_contest;
 pub use auth::{Authentication, GitHubAuthentication, GitHubUserResponse};
 
 use crate::server::ranking::{
-    get_ac_ranking, get_streak_ranking, get_users_ac_rank, get_users_streak_rank,
+    get_ac_ranking, get_streak_ranking, get_users_ac_rank, get_users_streak_rank, get_users_language_rank,
 };
 use auth::get_token;
-use language_count_ranking::get_users_language_rank;
 use middleware::LogMiddleware;
 use problem_list::{
     add_item, create_list, delete_item, delete_list, get_own_lists, get_single_list, update_item,
