+++ atcoder-problems-backend/src/server/ranking/mod.rs
@@ -1,6 +1,7 @@
 use crate::server::{AppData, CommonResponse};
 use serde::{Deserialize, Serialize};
 use sql_client::accepted_count::AcceptedCountClient;
+use sql_client::language_count::LanguageCountClient;
 use sql_client::streak::StreakClient;
 use std::ops::Range;
 use tide::Result;
@@ -123,4 +124,34 @@ pub(crate) async fn get_users_streak_rank<A>(
     };
     let rank = conn.get_streak_count_rank(count).await?;
     Ok(Some(UserRankResponse { count, rank }))
+}
+
+pub(crate) async fn get_users_language_rank<A>(
+    request: tide::Request<AppData<A>>,
+) -> Result<tide::Response> {
+    #[derive(Debug, Deserialize)]
+    struct Query {
+        user: String,
+    }
+    #[derive(Debug, Serialize)]
+    struct UsersLanguageResponse {
+        language: String,
+        count: i64,
+        rank: i64,
+    }
+    let conn = request.state().pg_pool.clone();
+    let query = request.query::<Query>()?;
+    let counts = conn.load_users_language_count(&query.user).await?;
+    let ranks = conn.load_users_language_count_rank(&query.user).await?;
+    let info = counts
+        .into_iter()
+        .zip(ranks)
+        .map(|(c, r)| UsersLanguageResponse {
+            language: c.simplified_language,
+            count: c.problem_count as i64,
+            rank: r.rank,
+        })
+        .collect::<Vec<_>>();
+    let response = tide::Response::json(&info)?;
+    Ok(response)
 }
